% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startDashboard.R
\name{dashboardServer}
\alias{dashboardServer}
\title{Create the dashboard's server}
\usage{
dashboardServer()
}
\value{
The UI definition of the app (for example, a call to fluidPage() with nested controls).
}
\description{
\code{\link[=dashboardServer]{dashboardServer()}} defines the logic of the dashboard and is to be passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}
as its \code{server} argument.
}
