% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startDashboard.R
\name{dashboardUi}
\alias{dashboardUi}
\title{Create the dashboard's UI}
\usage{
dashboardUi()
}
\value{
The UI definition of the app (for example, a call to fluidPage() with nested controls).
}
\description{
\code{\link[=dashboardUi]{dashboardUi()}} combines  all the UI elements used in the dashboard into
the \code{ui} object expected by \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.
}
