% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_viewer.R
\name{df_viewer_server}
\alias{df_viewer_server}
\title{Shiny Server for Data Frame Viewer}
\usage{
df_viewer_server(input, output, session)
}
\arguments{
\item{input}{input object from Shiny.}

\item{output}{output object from Shiny.}

\item{session}{session object from Shiny.}
}
\description{
Shiny Server for Data Frame Viewer
}
