% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_viewer.R
\docType{data}
\name{df_viewer_ui}
\alias{df_viewer_ui}
\alias{df_viewer_server}
\title{Shiny UI for data frame viewer application}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
df_viewer_ui

df_viewer_server(input, output, session)
}
\arguments{
\item{input}{input object from Shiny.}

\item{output}{output object from Shiny.}

\item{session}{session object from Shiny.}
}
\description{
This is a simple Shiny application to demonstrate the [runAppWithParams()] function. This will
display the structure and contents of any data frames listed in the `data_frames` list object.
}
\examples{
if (interactive()) { # Only run this example in interactive R sessions
data(mtcars)
data(faithful)
runAppWithParams(ui = ShinyDemo::df_viewer_ui,
                 server = ShinyDemo::df_viewer_server,
                 data_frames = list(mtcars = mtcars, faithful = faithful),
                 port = 2112)
}
}
\keyword{datasets}
