% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncomeDataSet.R
\name{getIncomeDataSet}
\alias{getIncomeDataSet}
\title{Get income data set}
\usage{
getIncomeDataSet()
}
\value{
Returns a \code{\link[base:data.frame]{base::data.frame()}}
}
\description{
Some formatting is applied as well.
}
\section{Description of the income data set}{
The data set is the \href{https://www.tidytuesday.com/}{Tidy Tuesdays}
\verb{Wealth and income over time}
\href{https://github.com/rfordatascience/tidytuesday/tree/master/data/2021/2021-02-09}{repository}
data set called \code{income_distribution.csv}. \code{income_distribution.csv} is
a cleaned version of the the raw data provided by the \href{https://apps.urban.org/features/wealth-inequality-charts/}{Urban Institute}
and the \href{https://www.census.gov/data/tables/time-series/demo/income-poverty/historical-income-households.html}{US Census}.
This file documents households by total money income, race, and Hispanic
origin of householder separated by year and income groups in USA.\tabular{lll}{
   variable \tab class \tab description \cr
   year \tab integer \tab Year \cr
   race \tab character \tab Racial Group \cr
   number \tab double \tab Number of households \cr
   income_median \tab integer \tab Income median \cr
   income_med_moe \tab integer \tab Income median margin of error \cr
   income_mean \tab integer \tab Income mean \cr
   income_mean_moe \tab integer \tab Income mean margin of error \cr
   income_bracket \tab character \tab Income bracket (9 total brackets between \verb{<$15,000} and \verb{>$200,000} \cr
   income_distribution \tab double \tab Income distribution as the percentage of each year/racial group - should add up to 100 for a specific year and race. \cr
}
}

