% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeVignette.R
\name{includeVignette}
\alias{includeVignette}
\title{Include a vignette within a Shiny app.}
\usage{
includeVignette(topic, package)
}
\arguments{
\item{topic}{a character string giving the name of the vignette to include.}

\item{package}{a character vector with the names of packages to search 
through, or NULL in which ‘all’ packages (as defined by argument all) 
are searched}
}
\description{
This will read in a vignette (currently HTML only) to be included within
a Shiny app.
}
