% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_all.R
\name{ls_all}
\alias{ls_all}
\title{List All Objects}
\usage{
ls_all()
}
\value{
a character vector with the name of all objects available.
}
\description{
This function returns a character vector of all objects available. Unlike
[ls()] this function will loop through all environments from the current
environment to \code{.GlobalEnv}. This will also verify that the object
is indeed available from the current environment using the [exists()]
function call.
}
