% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{make_hist}
\alias{make_hist}
\title{make_hist}
\usage{
make_hist(bins)
}
\arguments{
\item{bins}{numeric. Number of bins}
}
\value{
a hist object
}
\description{
make_hist
}
\author{
Simon Smart <simon.smart@cantab.net>
}
