% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIncomeData.R
\name{plotIncomeDistribution}
\alias{plotIncomeDistribution}
\alias{plotIncomeMedian}
\alias{plotPlaceHolder}
\title{Plot income data set}
\usage{
plotIncomeDistribution(year)

plotIncomeMedian(includeMarginOfError = FALSE)

plotPlaceHolder()
}
\arguments{
\item{year}{An integer between 1967 and 2019.}

\item{includeMarginOfError}{If \code{TRUE} margin of error is included. If \code{FALSE},
margin of error is not included.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Each functions plots a different aspect of the data set.
}
