% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderVignette.R
\name{renderVignette}
\alias{renderVignette}
\title{Render a vignette from source to include in a Shiny app.}
\usage{
renderVignette(topic, package, ...)
}
\arguments{
\item{topic}{a character string giving the name of the vignette to include.}

\item{package}{a character vector with the names of packages to search 
through, or NULL in which ‘all’ packages (as defined by argument all) 
are searched}

\item{...}{other parameters passed to [ShinyDemo::renderRmd()].}
}
\description{
This will render a vignette (currently Rmd only) to be included within
a Shiny app.
}
