% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{This is a wrapper to shiny::runApp to include parameters.}
\usage{
run_shiny_app(appDir, ui, server, ...)
}
\arguments{
\item{appDir}{the directory of the application to run.}

\item{ui}{the Shiny ui object.}

\item{server}{the Shiny server object.}

\item{...}{[shiny::runApp()] parameters, [shiny::shinyApp()] parameters,
or parameters to pass to the Shiny app.}
}
\description{
This function will run a Shiny app but will pass arbitrary parameters
(NAME = VALUE) through the \code{...} parameter to the application. This
is done by modifying the global environment. This function will attempt
to clean up any objects placed into the global environment on exit. If
objects exist prior to calling this function (i.e. \code{exists(OBJECT)}
returns TRUE) then the value will be reset to it's state prior to calling
\code{run_shiny_app}.
}
\examples{
\dontrun{
run_shiny_app(ui = ShinyDemo::df_viewer_ui,
              server = ShinyDemo::df_viewer_server,
              mtcars = mtcars)
}
}
