% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_demo.R
\name{shiny_demo}
\alias{shiny_demo}
\alias{print.shinyapplist}
\title{Find and run Shiny applications from R package}
\usage{
shiny_demo(
  topic,
  package,
  lib.loc = .libPaths(),
  verbose = getOption("verbose"),
  include.installed = FALSE,
  ...
)

\method{print}{shinyapplist}(x, ...)
}
\arguments{
\item{topic}{the topic/app which should be run.}

\item{package}{the package which contains the app to run. If `NULL` the 
first app with the given topic name will be run.}

\item{lib.loc}{a character vector of directory names of R libraries, or NULL.
The default value of NULL corresponds to all libraries currently known. If
the default is used, the loaded packages are searched before the libraries.}

\item{verbose}{a logical. If TRUE, additional diagnostics are printed.}

\item{include.installed}{search installed packages for Shiny applications. If `FALSE` only
loaded packages will be searched.}

\item{...}{other parameters passed to `print()`.}

\item{x}{results from [shiny_demo()].}
}
\value{
if `topic` is not specified this will return a data frame listing all the Shiny
  applications found.
}
\description{
`shiny_demo()` is a user-friendly interface to finding and running Shiny applications 
from R packages. For package developers, simply put Shiny apps in thec`inst/` directory in 
your package. This function will find any apps located there from loaded package.
}
\examples{
if(interactive()) {
library(ShinyDemo)
shiny_demo() # this should at least return the Shiny apps in this package
shiny_demo(topic = 'df_viewer', package = 'ShinyDemo')
}
}
