% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDemo-package.R
\docType{package}
\name{ShinyDemo-package}
\alias{ShinyDemo-package}
\title{Shiny Apps as Package Demos}
\description{
Shiny Package Demos
}
\author{
\email{jason@bryer.org}
}
\keyword{package}
\keyword{shiny}
