% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseIncomeData.R
\docType{package}
\name{visualiseIncomeData}
\alias{visualiseIncomeData}
\alias{visualiseIncomeDataMenuItem}
\alias{visualiseIncomeDataTabItem}
\alias{visualiseIncomeDataServer}
\title{Describe the extension in a few words}
\usage{
visualiseIncomeDataMenuItem(id)

visualiseIncomeDataTabItem(id)

visualiseIncomeDataServer(id)
}
\arguments{
\item{id}{The identifier string linking the \link{triplet} of functions together.}
}
\value{
These functions are to be used in \code{\link[=dashboardUi]{dashboardUi()}} and \code{\link[=dashboardServer]{dashboardServer()}} with the \emph{same} value of \code{id}.
\itemize{
\item \code{MenuItem(id)}: is called in \code{\link[=dashboardUi]{dashboardUi()}} to create an argument to \code{\link[shinydashboard:sidebarMenu]{shinydashboard::sidebarMenu()}}.
\item \code{TabItem(id)} : is called in \code{\link[=dashboardUi]{dashboardUi()}} to create an argument to \code{\link[shinydashboard:tabItems]{shinydashboard::tabItems()}}.
\item \code{Server(id)}  : is called in \code{\link[=dashboardServer]{dashboardServer()}}.
}

To know more about how to use these functions, read the documentation of the
\link{triplet}.
}
\description{
An extensive description of the extension.
}
\concept{extensions}
