% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseIncomeDataModule.R
\name{visualiseIncomeDataSelectionUi}
\alias{visualiseIncomeDataSelectionUi}
\alias{visualiseIncomeDataOptionUi}
\alias{visualiseIncomeDataPlotUi}
\alias{visualiseIncomeDataPlotServer}
\title{A module to visualise the income data}
\usage{
visualiseIncomeDataSelectionUi(id)

visualiseIncomeDataOptionUi(id)

visualiseIncomeDataPlotUi(id)

visualiseIncomeDataPlotServer(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}
}
\value{
\code{\link[=visualiseIncomeDataSelectionUi]{visualiseIncomeDataSelectionUi()}} returns a \code{\link[shiny:fluidPage]{shiny::fluidRow()}}.
}
\description{
This module as three front end elements for more flexibility in the layout.
}
\concept{\pkg{shiny} module functions.}
