% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{This is a wrapper to shiny::runApp to include parameters.}
\usage{
run_shiny_app(appDir, ui, server, ...)
}
\arguments{
\item{appDir}{the directory of the application to run.}

\item{ui}{}

\item{...}{[shiny::runApp()] parameters, [shiny::shinyApp()] parameters,
or parameters to pass to the Shiny app.}
}
\description{
This is a wrapper to shiny::runApp to include parameters.
}
