% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_demo.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{This is a wrapper to shiny::runApp to include parameters.}
\usage{
run_shiny_app(appDir, ...)
}
\arguments{
\item{appDir}{the directory of the application to run.}

\item{...}{[shiny::runApp] parameters or parameters to pass to the Shiny app.}
}
\description{
This is a wrapper to shiny::runApp to include parameters.
}
