% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDemo-package.R
\docType{package}
\name{ShinyDemo-package}
\alias{ShinyDemo}
\alias{ShinyDemo-package}
\title{Shiny Applications with Parameters}
\description{
Shiny Applications with Parameters
}
\details{
This package is designed to run Shiny apps (Chang, et al,
2021 https://shiny.rstudio.com> Wickham, 2021 isbn:9781492047384)
included in packages in the inst/shiny/ directory. In addition to
providing a standardized way of running apps from within packages, it
extends the typical process of running apps by allowing function
parameters to be passed to the application.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jbryer.github.io/ShinyDemo/}
  \item Report bugs at \url{https://github.com/jbryer/ShinyDemo/issues}
}

}
\author{
\strong{Maintainer}: Jason Bryer \email{jason@bryer.org} (\href{https://orcid.org/0000-0002-2454-0402}{ORCID})

}
\keyword{package}
\keyword{shiny}
