% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDemo-package.R
\docType{package}
\name{ShinyDemo-package}
\alias{ShinyDemo}
\alias{ShinyDemo-package}
\title{Shiny Apps as Package Demos}
\description{
Shiny Package Demos
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jbryer.github.io/ShinyDemo/}
  \item Report bugs at \url{https://github.com/jbryer/ShinyDemo/issues}
}

}
\author{
\email{jason@bryer.org}
}
\keyword{package}
\keyword{shiny}
